Fixed mutex issue
This is working on my macbook currently, will test on windows in 20 minutes. Should work there too, these changes were to fix scoping issues for nested locks on the gameState. We only need the gameState for reference so we can create a lock, pull what we need, and unlock it then continue our logic and lock the game as needed.